+++ conftest.py
@@ -2,15 +2,14 @@
 
 # third party
 import pytest
-from psycopg2.extras import RealDictCursor
 
 # This repo
-from pyiem.database import get_dbconn
+from pyiem.database import get_dbconnc
 
 
 @pytest.fixture()
 def dbcursor(database):
     """Yield a cursor for the given database."""
-    dbconn = get_dbconn(database)
-    yield dbconn.cursor(cursor_factory=RealDictCursor)
+    dbconn, cursor = get_dbconnc(database)
+    yield cursor
     dbconn.close()